🔥 chore: move src/api/routes directory to src/routes
